+++ frontend/src/components/MapToolbar.vue
@@ -2,7 +2,14 @@
 import { computed } from "vue";
 
 import { store } from "../store.js";
-import { cancelShape, finishShape, setMode } from "../actions.js";
+import {
+  cancelShape,
+  finishShape,
+  redoEdit,
+  setMode,
+  undoEdit,
+} from "../actions.js";
+import { redoButtonTitle, undoButtonTitle } from "../undo.js";
 import { cancelCropDraw, closeCropPolygon, startCropDraw } from "../map.js";
 
 const activeFeeds = computed(
@@ -46,6 +53,23 @@ const activeFeeds = computed(
       <button @click="finishShape">Finish</button>
       <button @click="cancelShape">Cancel</button>
     </template>
+
+    <button
+      class="mode"
+      :disabled="!store.undoLabel"
+      :title="undoButtonTitle(store.undoLabel)"
+      @click="undoEdit"
+    >
+      ↶ Undo
+    </button>
+    <button
+      class="mode"
+      :disabled="!store.redoLabel"
+      :title="redoButtonTitle(store.redoLabel)"
+      @click="redoEdit"
+    >
+      ↷ Redo
+    </button>
 
     <!-- Cropping: draw an area, then confirm in the panel below the map. -->
     <template v-if="!store.cropDrawing && !store.cropShape">